backend/wardrobe-management/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-auth
pytest
httpx
alembic
psycopg2-binary
requests